refactor: revert to inline classes — live editor errors on external file imports
The live editor raised import errors when importing from separate module files
(activity_log_service, pet_data_service, llm_service, external_api_service).
The OpenHome platform only loads main.py and does not add the ability directory
to sys.path, so any import from sibling files fails at runtime.

Changes:
- Inline ActivityLogService, PetDataService, LLMService, ExternalAPIService
  directly into main.py
- Restore missing is_hard_exit() method on LLMService
- Remove register_capability() classmethod (used blocked open() call);
  keep only #{{register capability}} tag per current template pattern
- Delete separate service files: activity_log_service.py, pet_data_service.py,
  llm_service.py, external_api_service.py
- Update tests/conftest.py and test_exit_detection.py to import from main
- Update README architecture diagram to reflect single-file structure
- All 220 tests pass
